modified model for user

+++ conferenceApp/accounts/views.py
@@ -15,36 +15,13 @@ from django.utils.encoding import force_str, force_bytes
 from django.contrib.auth.tokens import default_token_generator
 from allauth.socialaccount.models import SocialAccount
 from rest_framework_simplejwt.tokens import RefreshToken
-from .serializers import SignupSerializer, SigninSerializer, UserProfileSerializer, ResetPasswordSerializer, CompletePasswordResetSerializer
+from .serializers import CustomUserSerializer, SigninSerializer, UserProfileSerializer, ResetPasswordSerializer, CompletePasswordResetSerializer
+from django.contrib.auth import get_user_model
+from rest_framework import permissions
 
-
-
-@api_view(['GET'])
-# @permission_classes([IsAuthenticated])
-def home(request):
-    user = request.user
-
-    if user:
-        login(request, user, backend='allauth.account.auth_backends.AuthenticationBackend')
-
-        # Generate or retrieve the refresh token
-        refresh = RefreshToken.for_user(user)
-        access_token = str(refresh.access_token)
-
-        return JsonResponse({
-            "user_id": user.id,
-            "username": user.username,
-            "email": user.email,
-            "access_token": access_token,
-            
-        })
-
-    else:
-        return Response({'error': 'user not authenticated'}, status=status.HTTP_401_UNAUTHORIZED)
-    
 class SignupView(generics.CreateAPIView):
-    """Register user with email and password"""
-    serializer_class = SignupSerializer 
+    """Register user with email, password, and other details"""
+    serializer_class = CustomUserSerializer 
 
     def create(self, request, *args, **kwargs):
         response = super().create(request, *args, **kwargs)
@@ -55,7 +32,6 @@ class SignupView(generics.CreateAPIView):
             }, status=status.HTTP_201_CREATED)
         else:
             return response
-
 
 class SigninView(APIView):
     """Login a user with username and password"""
@@ -88,25 +64,36 @@ class SigninView(APIView):
             }, status=status.HTTP_200_OK)
         else:
             return Response({'error': 'Invalid credentials.'}, status=status.HTTP_401_UNAUTHORIZED)
+        
+
 class UserProfileView(generics.RetrieveUpdateDestroyAPIView):
     """Get user details, update, or delete a user"""
-    queryset = User.objects.all()  
-    serializer_class = UserProfileSerializer
-    permission_classes = [IsAuthenticated]
+    queryset = get_user_model().objects.all()  
+    serializer_class = CustomUserSerializer
+    permission_classes = [permissions.IsAuthenticated]
 
     def update(self, request, *args, **kwargs):
-        user = get_object_or_404(User, pk=request.user.id)
+        user = self.get_object()
         serializer = self.get_serializer(user, data=request.data, partial=True)
         serializer.is_valid(raise_exception=True)
+
+        # Check if the user has permission to update 'role' field
+        if not request.user.is_staff:
+            serializer.validated_data.pop('role', None)
+
         self.perform_update(serializer)
         return Response(serializer.data)
 
     def destroy(self, request, *args, **kwargs):
-        user = get_object_or_404(User, pk=request.user.id)
+        user = self.get_object()
+        
+        # Check if the user has permission to delete the user
+        if not request.user.is_staff and request.user != user:
+            return Response({'error': 'Permission denied.'}, status=status.HTTP_403_FORBIDDEN)
+
         user.delete()
         return Response({'message': 'User deleted successfully.'}, status=status.HTTP_204_NO_CONTENT)
     
-
 
 def send_reset_password_mail(email, resetlink):
     subject = 'Password reset request'
@@ -123,7 +110,7 @@ class ResetPasswordView(generics.CreateAPIView):
     def create(self, request, *args, **kwargs):
         email = request.data.get('email')
 
-        user = User.objects.filter(email=email).first() 
+        user = get_user_model().objects.filter(email=email).first() 
 
         if user is not None:
             # Generate a token for password reset
@@ -151,7 +138,7 @@ class CompletePasswordResetView(generics.CreateAPIView):
         
         uid = force_str(urlsafe_base64_decode(uid))  # Decode and convert to string
         
-        user = User.objects.filter(pk=uid).first()
+        user = get_user_model().objects.filter(pk=uid).first()
 
         if user is not None and default_token_generator.check_token(user, token):
             if new_password:
@@ -191,74 +178,3 @@ def get_user_details(request):
         # Add more user details specific to social accounts if needed
 
     return Response(response_data)
-
-
-
-# class GitHubSocialAuth(APIView):
-#     def get(self, request, code=None):
-#         if not code:
-#             return Response({"error": "Authorization code not provided."}, status=status.HTTP_400_BAD_REQUEST)
-
-#         try:
-#             # Exchange the authorization code for an access token
-#             adapter = GitHubOAuth2Adapter(request)
-#             client = OAuth2Client(
-#                 request,
-#                 os.getenv('GITHUB_CLIENT_ID'),
-#                 os.getenv('GITHUB_CLIENT_SECRET'),
-#                 'POST',  # Use 'POST' as the access token method
-#                 'https://github.com/login/oauth/access_token',  # GitHub's access token URL
-#                 'http://localhost:8000/accounts/github/login/callback/',  # Your actual callback URL
-#                 'user',  # Use 'user' scope to access basic user information
-#             )
-#             token = adapter.access_token(code, client)
-#             account = adapter.complete_login(request, app=None, token=token)
-#         except Exception as e:
-#             return Response({"error": str(e)}, status=status.HTTP_400_BAD_REQUEST)
-
-#         # Perform additional actions as needed, e.g., create a user or log in
-#         if account and account.user.is_authenticated:
-#             # Save user details to the User model
-#             save_user_details(account.user)
-
-#             # Log in the user
-#             login(request, account.user)
-
-#             # Generate or retrieve the authentication token
-#             token, created = Token.objects.get_or_create(user=account.user)
-
-#             # Return the user details and authentication token
-#             return Response({
-#                 "user_id": account.user.id,
-#                 "username": account.user.username,
-#                 "email": account.user.email,
-#                 "auth_token": token.key,  # Include the authentication token
-#                 # Add more user details as needed
-#             })
-
-#         # If the authentication process fails, return an error message
-#         return Response({"error": "Authentication failed."}, status=status.HTTP_400_BAD_REQUEST)
-
-# def save_user_details(user):
-#     # Extract user details from the social authentication provider
-#     social_account = SocialAccount.objects.get(user=user)
-#     extra_data = social_account.extra_data
-
-#     # Extract relevant information (customize based on your needs)
-#     username = extra_data.get('login', '')
-#     email = extra_data.get('email', '')
-#     name = extra_data.get('name', '')
-
-#     # Check if the user already exists based on email
-#     existing_user = User.objects.filter(email=email).first()
-
-#     if existing_user:
-#         # If the user exists, update relevant details
-#         existing_user.username = username
-#         existing_user.save()
-#     else:
-#         # If the user doesn't exist, create a new user
-#         new_user = User.objects.create_user(username=username, email=email)
-#         new_user.first_name = name
-#         new_user.save()
-
